refactor: add initialState service
